Provide requests DynamoDB table
Modify sceptre templates to create a DynamoDB table called "requests" with a
key attribute of "request_id".  An optional "ttl" attribute can be used to
automatically expire records after a certain point in time:

https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/TTL.html

These changes also create a new managed policy that allows DynamoDB access, and
adds the policy to the "Threat" role that's used by the lambda functions.