Asynchronous lambda API (proof of concept)
The /hello/{path_arg1} REST endpoint can be used to test synchronous vs.
asynchronous processing.  If the header parameter "async" is missing or is set
to "false", then the lambda is invoked synchronously, and the results are
returned as a JSON structure (status code 200).

If the header parameter "async" is set to "true", then this lambda will invoke
itself asynchronously (after setting the "async" parameter to "false" to avoid
an infinite loop).  The jobId is returned to the original caller (status code
202), and the secondary, synchronous invocation of the lambda will store its
results in the jobs table in DynamoDB.  The final output can be retrieved by
using the /job/{jobId} REST endpoint, specifying the provided jobId.