Update manager lambda (golang and python implementations)
This commit updates some of the boilerplate surrounding the manager lambda
function.  The sceptre templates build the golang version--the python version
is preserved at the moment for historical documentation and testing purposes.

The `update-lambda-golang.sh` script may be used while in development to build
and update the golang implementation of this lambda.