Add system lambdas to sceptre templates
The "manager" and "responseprocessor" lambdas are created and updated by the
sceptre templates.  Both are assumed to be golang implementations.  The SHA1
hash of the .go source file is used as the S3 key for the build, so that
sceptre and cloudformation can intelligently decide whether the function needs
to be updated.

The functions are currently built during each invocation, whether they've been
updated or not.