Add service lambdas to sceptre templates
The lambdas found under the apis/ directory are created and updated by the
sceptre templates.  Each subdirectory must contain a "lambda.json" file that
describes the parameters used to create the lambda, and a "build.sh" script
that's used to build the lambda.  The output of "build.sh" is expected to be a
ZIP file "function.zip" which is uploaded to S3.  The SHA1 hash of the .go or
.py source file is used as the S3 key for the build, so that sceptre and
cloudformation can intelligently decide whether the function needs to be
updated.

The functions are currently built during each invocation, whether they've been
updated or not.