Configure authorizer and manager lambdas to run on DX subnets
Certain environments (dev-private and dev) will now use sso.dev-gdcorp.tools as
the SSO endpoint, but this endpoint is only accessible via DirectConnect.  So
far, the authorizer and manager lambdas are the only lambdas that need access
to the dev SSO endpoint, so they need to be configured to run in DX-specific
subnets.

The SSO endpoint is available in Parameter Store as /ThreatTools/SSOHost, and
is also set in the SSO_HOST environment variable for these two lambdas.