Adjust SHA1 calculations for lambda functions to reduce deployment churn
SHA1 calculations are performed for each lambda function deployed with Service
Catalog.  These SHA1 values are used to determine whether or not a new version
of the lambda needs to be deployed.  This PR modifies the algorithm to take
into consideration the content of the resulting binary as opposed to the
generated zip file.  (The generated zip file would contain different content
each time the build was performed, due to timestamps being embedded into the
zip file.)  This should reduce the number of times that a lambda is deployed
when it hasn't really changed, while still taking into consideration changing
dependent libraries, etc.